refactor: renaming, use yup.InferType for form typing
- should prefer using schema even when validation is not used but
  resolved the issue with trying to use an undefined schema in useForm
  anyways
